Remove quic transport, move route into router
