+++ tunnel/options.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/micro/go-micro/v3/transport"
-	"github.com/micro/go-micro/v3/transport/quic"
+	"github.com/micro/go-micro/v3/transport/grpc"
 )
 
 var (
@@ -140,6 +140,6 @@ func DefaultOptions() Options {
 		Id:        uuid.New().String(),
 		Address:   DefaultAddress,
 		Token:     DefaultToken,
-		Transport: quic.NewTransport(),
+		Transport: grpc.NewTransport(),
 	}
 }
